Idk what I did but fuck it, something changed, desu.

+++ Welt/Controllers/FirstPersonCameraController.cs
@@ -62,7 +62,8 @@ namespace Welt.Controllers
 
             if (moveVector != Vector3.Zero)
             {
-                var rotationMatrix = Matrix.CreateRotationX(Camera.UpDownRotation) * Matrix.CreateRotationY(Camera.LeftRightRotation);
+                var rotationMatrix = Matrix.CreateRotationX(Camera.UpDownRotation)*
+                                     Matrix.CreateRotationY(Camera.LeftRightRotation);
                 var rotatedVector = Vector3.Transform(moveVector, rotationMatrix);
                 Camera.Position += rotatedVector * MOVEMENTSPEED;
             }
@@ -70,6 +71,7 @@ namespace Welt.Controllers
         #endregion
 
         #region Update
+
         public void Update(GameTime gameTime)
         {
             var currentMouseState = Mouse.GetState();
@@ -79,32 +81,34 @@ namespace Welt.Controllers
 
             if (mouseDx != 0)
             {
-                Camera.LeftRightRotation -= ROTATIONSPEED * (mouseDx / 50);
+                Camera.LeftRightRotation -= ROTATIONSPEED*(mouseDx/50);
             }
             if (mouseDy != 0)
             {
-                Camera.UpDownRotation -= ROTATIONSPEED * (mouseDy / 50);
+                Camera.UpDownRotation -= ROTATIONSPEED*(mouseDy/50);
 
                 // Locking camera rotation vertically between +/- 180 degrees
-                var newPosition = Camera.UpDownRotation - ROTATIONSPEED * (mouseDy / 50);  
-                if (newPosition < -1.55f)  
-                    newPosition = -1.55f;  
-                else if (newPosition > 1.55f)  
-                    newPosition = 1.55f;  
-                Camera.UpDownRotation = newPosition;  
+                var newPosition = Camera.UpDownRotation - ROTATIONSPEED*(mouseDy/50);
+                if (newPosition < -1.55f)
+                    newPosition = -1.55f;
+                else if (newPosition > 1.55f)
+                    newPosition = 1.55f;
+                Camera.UpDownRotation = newPosition;
                 // End of locking
             }
 
             //camera.LeftRightRotation -= GamePad.GetState(Game.ActivePlayerIndex).ThumbSticks.Right.X / 20;
             //camera.UpDownRotation += GamePad.GetState(Game.ActivePlayerIndex).ThumbSticks.Right.Y / 20;
 
-            _mMouseMoveState = new MouseState(Camera.Viewport.Width / 2,
-                    Camera.Viewport.Height / 2,
-                    0, ButtonState.Released, ButtonState.Released, ButtonState.Released, ButtonState.Released, ButtonState.Released);
+            _mMouseMoveState = new MouseState(Camera.Viewport.Width/2,
+                Camera.Viewport.Height/2,
+                0, ButtonState.Released, ButtonState.Released, ButtonState.Released, ButtonState.Released,
+                ButtonState.Released);
 
-            Mouse.SetPosition((int)_mMouseMoveState.X, (int)_mMouseMoveState.Y);
+            Mouse.SetPosition(_mMouseMoveState.X, _mMouseMoveState.Y);
             _mMouseState = Mouse.GetState();
         }
+
         #endregion
 
     }
